+++ CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add new client credential flow example in the Get-Entra cmdlet
 - Add new client credential flow example in the README
+- Add new authorization code flow example in the Get-Entra cmdlet
+- Add new authorization code flow example in the README
 
 ### Fixed
 
 - Get-EntraToken -ClientCredentialFlowWithSecret output a non necessary line in the output. Remove it.
+- Get-EntraToken -PublicAuthorizationCodeFlow output a non necessary line in the output. Remove it.
 
 ## [0.0.1] - 2023-09-23